robust_strategy_discovery/src/requirements.txt
absl-py==0.8.0
astor==0.8.0
attrs==19.1.0
autopep8==1.4.4
backcall==0.1.0
backports.functools-lru-cache==1.4
bleach==3.1.0
certifi==2019.6.16
chardet==3.0.4
cloudpickle==0.5.2
contexttimer==0.3.3
cycler==0.10.0
Cython==0.27.3
decorator==4.1.2
defusedxml==0.6.0
entrypoints==0.3
future==0.16.0
gast==0.3.1
graphviz==0.12
grpcio==1.23.0
gym==0.9.4
h5py==2.10.0
idna==2.6
ipykernel==5.1.2
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.11.1
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyter-kernel-gateway==2.4.0
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.17.0
pandas==0.25.1
pandocfilters==1.4.2
parso==0.1.1
patsy==0.5.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==6.1.0
pkg-resources==0.0.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.1
ptyprocess==0.5.2
pycodestyle==2.5.0
PyContracts==1.8.0
pyglet==1.3.0
Pygments==2.2.0
pygmo==2.6
pyparsing==2.4.2
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.4
requests==2.18.4
scikit-garden==0.1.3
scikit-learn==0.19.1
scikit-optimize==0.4
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.12.0
statsmodels==0.8.0
tensorboard==1.12.2
tensorflow-gpu==1.12.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
toolz==0.9.0
torch==1.2.0
torchvision==0.4.0
tornado==6.0.3
tqdm==4.35.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.6
widgetsnbextension==3.5.1
